Fix 5 critical issues: x64 PE detection, offset truncation, header limit, buffer overflow, code deduplication
- Fix get_nt_hrds64(): was checking I386 instead of AMD64 (const-correct)
- Fix 64-bit offset truncation: get_section_hdr() now takes DWORD64 instead of int
- Increase kMaxOffset from 1024 to 65536 for valid PE files with large e_lfanew
- Fix Unicode case-insensitive buffer overflow in search_chunk()
- Deduplicate algorithms: remove ~100 lines duplicated between algo.h and search_helper.cpp
- Fix winnt_mock.h: ULONG typedef before use, ULONGLONG above first usage
- Fix algo.h: const-cast on section pointer, fix section name string literals
- Fix tests/CMakeLists.txt: add PEFindC1 include path for algo.h
- Add cross-platform sprintf_s/snprintf compatibility in search_helper.cpp

+++ PEFindC1/pe_hdrs_helper.h
@@ -7,5 +7,5 @@ IMAGE_NT_HEADERS64* get_nt_hrds64(const BYTE* pe_buffer);
 bool is64bit(const BYTE* pe_buffer);
 IMAGE_DATA_DIRECTORY* get_pe_directory(const BYTE* pe_buffer, DWORD dir_id);
 ULONGLONG get_module_base(const BYTE* pe_buffer);
-PIMAGE_SECTION_HEADER get_section_hdr(const BYTE* payload, const size_t buffer_size, int globalOffset, int &sectionIndex);
+PIMAGE_SECTION_HEADER get_section_hdr(const BYTE* payload, const size_t buffer_size, DWORD64 globalOffset, int &sectionIndex);
 BOOL checkPE(const BYTE* buf);
